validating employees in controller

+++ ProiectSalariati/src/main/java/salariati/controller/EmployeeController.java
@@ -6,35 +6,39 @@ import salariati.model.Employee;
 import salariati.repository.interfaces.EmployeeRepositoryInterface;
 
 public class EmployeeController {
-	
-	private EmployeeRepositoryInterface employeeRepository;
-	
-	public EmployeeController(EmployeeRepositoryInterface employeeRepository) {
-		this.employeeRepository = employeeRepository;
-	}
-	
-	public void addEmployee(Employee employee) {
-		employeeRepository.addEmployee(employee);
-	}
-	
-	public List<Employee> getEmployeesList() {
-		return employeeRepository.getEmployeeList();
-	}
-	
-	public void modifyEmployee(Employee oldEmployee, Employee newEmployee) {
-		employeeRepository.modifyEmployee(oldEmployee, newEmployee);
-	}
 
-	public void deleteEmployee(Employee employee) {
-		employeeRepository.deleteEmployee(employee);
-	}
+    private EmployeeRepositoryInterface employeeRepository;
+    private EmployeeValidator employeeValidator = new EmployeeValidator();
 
-	public List<Employee> getEmployeesByAgeAsc(){
-		return employeeRepository.getEmployeesByAgeAsc();
-	}
+    public EmployeeController(EmployeeRepositoryInterface employeeRepository) {
+        this.employeeRepository = employeeRepository;
+    }
 
-	public List<Employee> getEmployeesBySalaryDesc(){
-		return employeeRepository.getEmployeesBySalaryDesc();
-	}
-	
+    public void addEmployee(Employee employee) {
+        if (employeeValidator.isValid(employee))
+            employeeRepository.addEmployee(employee);
+    }
+
+    public List<Employee> getEmployeesList() {
+        return employeeRepository.getEmployeeList();
+    }
+
+    public void modifyEmployee(Employee oldEmployee, Employee newEmployee) {
+        if (employeeValidator.isValid(newEmployee))
+            employeeRepository.modifyEmployee(oldEmployee, newEmployee);
+    }
+
+    public void deleteEmployee(Employee employee) {
+        if (employeeValidator.isValid(employee))
+            employeeRepository.deleteEmployee(employee);
+    }
+
+    public List<Employee> getEmployeesByAgeAsc() {
+        return employeeRepository.getEmployeesByAgeAsc();
+    }
+
+    public List<Employee> getEmployeesBySalaryDesc() {
+        return employeeRepository.getEmployeesBySalaryDesc();
+    }
+
 }
